LP-1506 Moving latest_course_reruns to helpers & updating tests
